powerpc: Split systemcfg struct definitions out from vdso
The systemcfg data has nothing to do anymore with the vdso.
Split it into a dedicated header file.
